% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epaSlimFooter.R
\name{epaSlimFooter}
\alias{epaSlimFooter}
\title{epaSlimFooter Function}
\usage{
epaSlimFooter(inputId, appVersion = "v0.0.0", appPublished = "local")
}
\arguments{
\item{inputId}{The id of the shiny object.}

\item{appVersion}{The application version}

\item{appPublished}{The application publish date}
}
\description{
Used in the UI functions of R Shiny apps
}
