% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epaComponents.R, R/epaSlimHeader.R
\name{epaSlimHeader}
\alias{epaSlimHeader}
\title{epaSlimHeader Function}
\usage{
epaSlimHeader(inputId, appEnvironment = "dev")

epaSlimHeader(inputId, appEnvironment = "dev")
}
\arguments{
\item{inputId}{The id of the shiny object.}

\item{appEnvironment}{The application environment. Supplying 'prod' will omit dev/testing banner.}
}
\description{
Used in the UI functions of R Shiny apps

Used in the UI functions of R Shiny apps
}
